Moved DepotType to the hudson.plugins.perforce.utils namespace
